+++ packages/alert/src/AlertBase.tsx
@@ -1,9 +1,9 @@
 import Button from '@littlespoon/button'
 import Link from '@littlespoon/link'
+import colors from '@littlespoon/theme/lib/colors'
 import { p3, p4 } from '@littlespoon/theme/lib/fonts/paragraph'
 import { family, weight } from '@littlespoon/theme/lib/fonts/primary'
 import { rem } from '@littlespoon/theme/lib/utils'
-import colors from '@littlespoon/theme/src/colors'
 import styled from 'styled-components'
 
 import type { AlertProps } from './Alert'