+++ packages/storybook/src/stories/icons/SubtractIcon.stories.tsx
@@ -1,0 +1,42 @@
+import SubtractIcon from '@littlespoon/icons/src/SubtractIcon'
+import { ComponentMeta, ComponentStory } from '@storybook/react'
+
+export default {
+  title: 'Design System/Icons/SubtractIcon',
+  component: SubtractIcon,
+} as ComponentMeta<typeof SubtractIcon>
+
+const Template: ComponentStory<typeof SubtractIcon> = (args) => <SubtractIcon {...args} />
+
+export const Default = Template.bind({})
+Default.args = {}
+
+export const ExtraSmall = Template.bind({})
+ExtraSmall.args = {
+  size: 'xsmall',
+}
+
+export const Small = Template.bind({})
+Small.args = {
+  size: 'small',
+}
+
+export const Medium = Template.bind({})
+Medium.args = {
+  size: 'medium',
+}
+
+export const Large = Template.bind({})
+Large.args = {
+  size: 'large',
+}
+
+export const Fill = Template.bind({})
+Fill.args = {
+  fill: 'skyblue',
+}
+
+export const Stroke = Template.bind({})
+Stroke.args = {
+  stroke: 'tomato',
+}