+++ packages/storybook/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.30.1](https://github.com/little-spoon-dev/design-system/compare/storybook@1.30.0...storybook@1.30.1) (2022-08-30)
+
+**Note:** Version bump only for package storybook
+
+# [1.30.0](https://github.com/little-spoon-dev/design-system/compare/storybook@1.29.0...storybook@1.30.0) (2022-08-25)
+
+### Features
+
+- **icons:** create subtract icon ([6e26f75](https://github.com/little-spoon-dev/design-system/commit/6e26f75ae7eb5a81b0232d3a63a01f38b66fedfd))
+
 # [1.29.0](https://github.com/little-spoon-dev/design-system/compare/storybook@1.28.0...storybook@1.29.0) (2022-08-22)
 
 ### Bug Fixes